refactor(animation): group animation code into dedicated module tree
Move existing animation logic into `crates/halley-wl/src/animation/`,
removing cross-cutting concerns from render, input, and top-level modules.

Split functionality into focused components:
- shared curves
- node-state animation
- surface scaling helpers
- node movement animation

Update imports to preserve existing behavior.

This lays the groundwork for a more structured animation system,
enabling separate window, layer, node, and cluster animation paths
without reintroducing coupling across unrelated subsystems.
