+++ crates/halley-wl/src/input/mod.rs
@@ -4,7 +4,6 @@ use std::sync::OnceLock;
 mod input_events;
 mod input_utils;
 mod key_actions;
-mod move_anim;
 mod pointer_button;
 mod pointer_focus;
 mod pointer_motion;
@@ -12,7 +11,6 @@ mod resize_helpers;
 
 pub(crate) use input_events::{BackendInputEventData, handle_backend_input_event};
 pub(crate) use key_actions::spawn_command;
-pub(crate) use move_anim::advance_node_move_anim;
 pub(crate) use resize_helpers::{active_node_screen_rect, active_resize_geometry_screen};
 
 pub(crate) fn pointer_map_debug_enabled() -> bool {
