+++ crates/halley-wl/src/render/mod.rs
@@ -2,7 +2,6 @@ use glam::{Vec2, Vec4};
 use halley_core::field::Field;
 use halley_core::viewport::{FocusRing, FocusZone, Viewport};
 
-mod anim_utils;
 mod cursor_render;
 mod cursor_theme;
 mod frame_render;
@@ -11,9 +10,9 @@ mod node_render;
 mod offscreen;
 mod render_utils;
 
-pub(crate) use anim_utils::{active_surface_render_scale, ease_in_out_cubic};
 pub(crate) use frame_render::{draw_debug_frame, draw_debug_frame_to_target};
 pub(crate) use render_utils::{node_marker_bounds, node_marker_metrics, world_to_screen};
+pub(crate) use render_utils::preview_proxy_size;
 
 #[derive(Clone, Copy, Debug)]
 pub struct DebugPalette {
