+++ crates/halley-wl/src/state/runtime_state.rs
@@ -4,7 +4,7 @@ use halley_config::RuntimeTuning;
 
 use eventline::debug;
 
-use crate::anim::{AnimSpec, AnimStyle};
+use crate::animation::{AnimSpec, AnimStyle};
 use crate::render::{DebugScene, build_debug_scene};
 
 impl HalleyWlState {
