+++ crates/halley-wl/src/wm/mod.rs
@@ -13,7 +13,7 @@ use smithay::reexports::wayland_server::{
 };
 
 use crate::activity::{CommitActivity, VisualState};
-use crate::anim::{AnimSpec, AnimStyle};
+use crate::animation::{AnimSpec, AnimStyle};
 use crate::render::{DebugScene, build_debug_scene};
 use crate::state::HalleyWlState;
